+++ tests/single/test_02_auth_get.py
@@ -1,4 +1,4 @@
-def test_get_flow(client):
+def test_login(client):
     resp = client.post("/login", data={
         "username": "admin",
         "password": "admin123"
@@ -6,22 +6,27 @@ def test_get_flow(client):
 
     assert resp.status_code == 302
 
+def test_dashboard(client):
     dashboard = client.get("/dashboard")
     assert dashboard.status_code == 200
     assert b"your blog" in dashboard.data
 
+def test_edit(client):
     edit = client.get("/dashboard/edit/default")
     assert edit.status_code == 200
     assert b"Title:" in edit.data
 
+def test_post(client):
     new_post = client.get("/dashboard/post/default")
     assert new_post.status_code == 200
     assert b"Include author" in new_post.data
 
+def test_edit_post(client):
     edit_post = client.get("/dashboard/edit/default/test")
     assert edit_post.status_code == 200
     assert b"Include author" in edit_post.data
 
+def test_stats(client):
     stats = client.get("/dashboard/get_stats/1/1/24")
     assert stats.status_code == 200
     assert stats.content_type == "application/json"